requirements.txt
aiopg==1.4.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
channels_postgres==1.0.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.1
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
idna==3.6
Levenshtein==0.25.1
lxml==5.2.2
msgpack==1.0.8
numpy==2.1.0
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.9.4
redis==5.0.8
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.53
sqlparse==0.4.4
sslcommerz-lib==1.0
svglib==1.5.1
tinycss2==1.3.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.1
uvicorn==0.30.5
webencodings==0.5.1
whitenoise==6.6.0
wsproto==1.2.0
xhtml2pdf==0.2.16
zopfli==0.2.3
